remove ArrownDown component and add button with text and arrowDownCircle
